study session - button disabled fix, showing pomodoro for regular study type...

See merge request ateamindia/jochi-phase2/frontend!41

+++ components/user/UserTeacherSidebar.vue
@@ -693,11 +693,16 @@ export default {
       await this.markAsRead({ notificationId: id });
     },
     async GetLogout() {
-      await this.getLogout({
-        auth_token: localStorage.getItem("token").replace("Bearer ", ""),
-      });
-      window.localStorage.clear();
-      window.location.href = GG4L_REDIRECT_URL + FRONTEND_BASE_URL;
+      let confirm = window.confirm("You will be logged out now");
+      if (confirm) {
+        await this.getLogout({
+          auth_token: localStorage.getItem("token").replace("Bearer ", ""),
+        });
+        window.localStorage.clear();
+        window.location.href = GG4L_REDIRECT_URL + FRONTEND_BASE_URL;
+      } else {
+        return;
+      }
     },
     clickableIcon() {
       let element = document.getElementById("clickableId");
